getScore() actually gets your score
But, my hard work didn't exactly pay off.  I really wanted this to be procedural however I can't see a way for it to work.  The thing is, you for some reason can't return the value from the getScore() function so that it can be used inline.  This is due to the time it takes for the request to be handled by the server and sent back.  The app continues to run however after the request is posted and so the return value at the very end returns the previous global variable value instead of the one received from the request.